+++ packages/live/src/fiber.ts
@@ -107,7 +107,7 @@ export const makeFiber = <F extends ArrowFunction>(
     yeeted, quotes, quote: null, unquote, context,
     state: null, pointer: 0, version: null, memo: null, runs: 0,
     mount: null, mounts: null, next: null, order: null, lookup: null,
-    type: null, id, by,
+    type: null, id, by, key,
   } as LiveFiber<F>;
 
   self.bound = bind(f, self) as any as F;
@@ -140,6 +140,7 @@ export const makeSubFiber = <F extends ArrowFunction>(
 export const makeNextFiber = <F extends ArrowFunction>(
   fiber: LiveFiber<F>,
   Next: LiveFunction<any>,
+  key?: Key,
   prefix: string = 'Next',
   reyeet?: boolean,
   name?: string,
@@ -150,8 +151,9 @@ export const makeNextFiber = <F extends ArrowFunction>(
   if (name === prefix) name = '…';
   Next.displayName = `${prefix}(${name})`;
 
-  const nextFiber = makeSubFiber(fiber, use(Next), fiber.id, 1);
+  const nextFiber = makeSubFiber(fiber, use(Next), fiber.id, 1); // path hardcoded to 1 to go second
   (nextFiber.f as any).isLiveContinuation = (fiber.f as any).isLiveBuiltin;
+  nextFiber.key = key; // key set to allow remounting
 
   // Adopt existing yeet context
   // which will be overwritten on the original fiber.
@@ -342,11 +344,13 @@ export const updateFiber = <F extends ArrowFunction>(
     return;
   }
 
-  // If fiber type changed, remount everything
-  if (fiber.type && fiber.type !== fiberType) disposeFiberState(fiber);
-  fiber.type = fiberType as any;
-
   const callArgs = call?.args ?? EMPTY_ARRAY;
+  const callKey = call?.key;
+
+  // If fiber type changed, remount everything
+  if (fiber.next && callKey != null && fiber.next.key !== callKey) disposeFiberState(fiber);
+  else if (fiber.type && fiber.type !== fiberType) disposeFiberState(fiber);
+  fiber.type = fiberType as any;
 
   // Reconcile literal array
   if (isArray) {
@@ -362,22 +366,22 @@ export const updateFiber = <F extends ArrowFunction>(
   // Map reduce
   else if (fiberType === MAP_REDUCE) {
     const [calls, map, reduce, then, fallback] = callArgs;
-    mapReduceFiberCalls(fiber, calls, map, reduce, then, fallback);
+    mapReduceFiberCalls(fiber, calls, map, reduce, then, fallback, callKey);
   }
   // Gather reduce
   else if (fiberType === GATHER) {
     const [calls, then, fallback] = callArgs;
-    gatherFiberCalls(fiber, calls, then, fallback);
+    gatherFiberCalls(fiber, calls, then, fallback, callKey);
   }
   // Multi-gather reduce
   else if (fiberType === MULTI_GATHER) {
     const [calls, then, fallback] = callArgs;
-    multiGatherFiberCalls(fiber, calls, then, fallback);
+    multiGatherFiberCalls(fiber, calls, then, fallback, callKey);
   }
   // Fence gathered reduction
   else if (fiberType === FENCE) {
     const [calls, then, fallback] = callArgs;
-    fenceFiberCalls(fiber, calls, then, fallback);
+    fenceFiberCalls(fiber, calls, then, fallback, callKey);
   }
   // Signal quoted reduction directly
   else if (fiberType === SIGNAL) {
@@ -647,6 +651,7 @@ export const reconcileFiberOrder = <F extends ArrowFunction>(
 // Make a reconciling tail for a fiber. Used to fix order in case of rekeying.
 export const makeResolveFiber = <F extends ArrowFunction>(
   fiber: LiveFiber<F>,
+  key?: Key,
   name: string = 'Resolve',
 ) => {
   // Incrementally reconciled. Pre-initialize these.
@@ -659,7 +664,7 @@ export const makeResolveFiber = <F extends ArrowFunction>(
     reconcileFiberOrder(fiber);
   };
 
-  return makeNextFiber(fiber, Resume, name);
+  return makeNextFiber(fiber, Resume, 1, name);
 }
 
 const toArray = <T>(x: T | T[] | undefined): T[] => Array.isArray(x) ? x : x != null ? [x] : [];
@@ -674,9 +679,10 @@ export const mapReduceFiberCalls = <F extends ArrowFunction, R, T>(
   reducer: (a: R, b: R) => R,
   next?: LiveFunction<any>,
   fallback?: R | typeof SUSPEND,
+  key?: Key,
 ) => {
   const gather = reduceFiberValues(reducer);
-  return mountFiberReduction(fiber, calls, mapper, gather, next, fallback);
+  return mountFiberReduction(fiber, calls, mapper, gather, next, fallback, key);
 }
 
 // Gather-reduce a fiber
@@ -685,8 +691,9 @@ export const gatherFiberCalls = <F extends ArrowFunction, T>(
   calls: LiveElement,
   next?: LiveFunction<any>,
   fallback: (T[] | typeof SUSPEND) = NO_ARRAY,
+  key?: Key,
 ) => {
-  return mountFiberReduction(fiber, calls, undefined, gatherFiberValues, next, fallback);
+  return mountFiberReduction(fiber, calls, undefined, gatherFiberValues, next, fallback, key);
 }
 
 // Multi-gather-reduce a fiber
@@ -695,8 +702,9 @@ export const multiGatherFiberCalls = <F extends ArrowFunction, T>(
   calls: LiveElement,
   next?: LiveFunction<any>,
   fallback: (Record<string, T[]> | typeof SUSPEND) = NO_RECORD,
+  key?: Key,
 ) => {
-  return mountFiberReduction(fiber, calls, undefined, multiGatherFiberValues, next, fallback);
+  return mountFiberReduction(fiber, calls, undefined, multiGatherFiberValues, next, fallback, key);
 }
 
 // Fence a fiber reduction
@@ -705,10 +713,11 @@ export const fenceFiberCalls = <F extends ArrowFunction, T>(
   calls: LiveElement,
   next?: LiveFunction<any>,
   fallback?: T | typeof SUSPEND,
+  key?: Key,
 ) => {
   const {yeeted} = fiber;
   const gather = yeeted?.gather ?? NOP;
-  return mountFiberReduction(fiber, calls, undefined, gather, next, fallback);
+  return mountFiberReduction(fiber, calls, undefined, gather, next, fallback, key);
 }
 
 // Reduce yeeted values on a tree of fibers (values have already been mapped on emit)
@@ -932,10 +941,11 @@ export const mountFiberReduction = <F extends ArrowFunction, R, T>(
   gather: FiberGather<R>,
   Next?: LiveFunction<any>,
   fallback?: R,
+  key?: Key,
 ) => {
   if (!fiber.next) {
     const Resume = makeFiberReduction(fiber, gather, fallback);
-    fiber.next = makeNextFiber(fiber, Resume, 'Resume', true);
+    fiber.next = makeNextFiber(fiber, Resume, key, 'Resume', true);
     fiber.yeeted = makeYeetState(fiber, fiber.next, gather, mapper);
     fiber.path = [...fiber.path, 0];
   }
@@ -1162,7 +1172,7 @@ export const captureFiber = <F extends ArrowFunction>(
     };
 
     const Resume = makeFiberReduction(fiber, reduction);
-    fiber.next = makeNextFiber(fiber, Resume, 'Resume');
+    fiber.next = makeNextFiber(fiber, Resume, 1, 'Resume');
 
     fiber.context = makeContextState(fiber, fiber.context, fiber.next, capture, registry);
     fiber.path = [...fiber.path, 0];
@@ -1173,7 +1183,7 @@ export const captureFiber = <F extends ArrowFunction>(
   mountFiberContinuation(fiber, use(fiber.next.f, then));
 }
 
-// Provide a value for a context on a fiber
+// Mount a reconciler on a fiber
 export const reconcileFiber = <F extends ArrowFunction>(
   fiber: LiveFiber<F>,
 ) => {
@@ -1188,7 +1198,7 @@ export const reconcileFiber = <F extends ArrowFunction>(
     if (!reconciler.reconciler) throw new Error(`'${reconciler.displayName}' is not a reconciler`);
 
     // Dummy fiber to act as new tree root, never called directly
-    const next = fiber.next = makeNextFiber(fiber, () => { throw new Error(); }, 'Root', false, reconciler.displayName);
+    const next = fiber.next = makeNextFiber(fiber, () => { throw new Error(); }, 1, 'Root', false, reconciler.displayName);
     next.quotes = quotes;
 
     quotes = fiber.quotes = new Map(quotes);