+++ packages/workbench/src/helpers/index.ts
@@ -1,5 +1,6 @@
 export * from './aabb-helper';
 export * from './axis-helper';
+export * from './loading-spinner';
 export * from './point-helper';
 export * from './vector-helper';